Refactor lib module
